move bathymety class to new file
